hello, an article on the relationship between FLOSS project and compagnies' business model
Abstract
In this article, we propose a theory that explains how Free/Libre Open Software (FLOSS) projects work and how companies rely on these FLOSS projects to develop their commercial offers, what we refer to as their “open-source” business model(s). This article builds on and refines the studies of the FLOSS organization by connecting two interrelated aspects: (1) how this organization evolves over time, in order to (2) better understand the value that users create and capture at each moment of a FLOSS project, with a particular focus on open-source companies, which are specific users who do business based on the software created by the FLOSS project. We describe these models and show that the open-source business models of companies are based on contributing to FLOSS projects in order to be able to provide “3A” services (assurance, adaptation, and assistance or support for use) that are complementary to the access to the software. Providing these services requires participation in the FLOSS project, which provides the project with the resources to operate.
This work can help the software engineering community by showing how FLOSS evaluation tools can be improved by taking into account the maturity of the solution, the strategic need of the target user, and the complementary open-source offers that exist.